Move to proposed structure
